commons/utils/job-store-service-connector/src/main/java/dk/dbc/dataio/commons/utils/jobstore/JobStoreServiceConnector.java
package dk.dbc.dataio.commons.utils.jobstore;

import dk.dbc.dataio.commons.types.JobInfo;
import dk.dbc.dataio.commons.types.JobSpecification;
import dk.dbc.dataio.commons.utils.invariant.InvariantUtil;

import javax.ws.rs.client.Client;

/**
 * JobStoreServiceConnector - dataIO job-store REST service client.
 * <p>
 * To use this class, you construct an instance, specifying a web resources client as well as
 * a base URL for the job-store service endpoint you will be communicating with.
 * </p>
 * <p>
 * This class is thread safe, as long as the given web resources client remains thread safe.
 * </p>
 */
public class JobStoreServiceConnector {
    private final Client httpClient;
    private final String baseUrl;

    /**
     * Class constructor
     *
     * @param httpClient web resources client
     * @param baseUrl base URL for job-store service endpoint
     *
     * @throws NullPointerException if given null-valued argument
     * @throws IllegalArgumentException if given empty-valued {@code baseUrl} argument
     */
    public JobStoreServiceConnector(Client httpClient, String baseUrl) throws NullPointerException, IllegalArgumentException {
        this.httpClient = InvariantUtil.checkNotNullOrThrow(httpClient, "httpClient");
        this.baseUrl = InvariantUtil.checkNotNullNotEmptyOrThrow(baseUrl, "baseUrl");
    }

    /**
     *
     * @param jobSpecification
     * @return
     * @throws NullPointerException if given null-valued argument
     * @throws JobStoreServiceConnectorException
     */
    public JobInfo createJob(JobSpecification jobSpecification) throws NullPointerException, JobStoreServiceConnectorException {
        InvariantUtil.checkNotNullOrThrow(jobSpecification, "jobSpecification");
        return null;
    }

    public Client getHttpClient() {
        return httpClient;
    }

    public String getBaseUrl() {
        return baseUrl;
    }
}
